VIRO-583, VIRO-584: Initial Commit Android Project Workspace Setup
Initial Project setup. Added directories:
- Android: Contains all Viro java bridge files
- ios: Contains all Viro objective C bridge files
- components: Contains all Viro javascript files
- Examples: Includes an ios and an Android Viro Sample project
that uses the controls.

The project is setup in a way that follows the build flow:
1) If building the Viro Renderer, copy the AAR output of that
build and import into the react-viro/android library.
2) If changing the Viro Java bridge files, simply build it. And
then build the Examples/Android project - this should automatically
pull in all required files.

This change also includes the base skeleton layout for Viro and how the
creation of the ViroRenderer through custom react components.


Former-commit-id: fbbdcb6f167d88b55fa2933a457a467f3e0b592e