Basic Bridge JNI Framework
First pass at creating UI View controls through the react bridge, it's
managers and our scene navigators, so as to produce the Render
demo (a 360 photo of Nigara Falls an a box). Tested and verified
that all controls work from the bridge - jni - renderer and that
scene rendered as expected.

- Created a base viewgroup class (VRT Group) that contains all adding/
removing of nodes. It also contains a node from which to add / remove.
This group extends from FrameLayout such that React's ViewGroupManager
can create it. Scene extends from VRTGroup to allow the adding of
objects to a scene.
- Created a base UI Control class (VRTControl) that all Viro ui visible
controls will inherit from. This class will extend view and contains a node.
These nodes are then added to the main scene's node.
- Created Box and Box Manager that inherits from VRTControl and have
the ability to add geometry to a given node through jni..
- SceneNavigator is now a viewgroup that add Scene views, and points the
manager to a given scene.


Former-commit-id: 0ee0b590ee468835e0f43c2cf6a9d1445bb621a7