VIRO-433 [bridge]: node-based cameras
Integrates cameras with nodes. Cameras can be
added to nodes, and then a specific node can be
chosen to be the "point of view" for the scene.
This also enables cameras to be animated.

This change includes all bridge changes. Some notes:

1. The camera to use is that which has the active prop
set to true. The react bridge uses the React context
to find all the cameras in a subtree, and selects one
with active=true. If multiple cameras have active=true,
behavior is undefined.

2. Each time a camera is mounted or unmounted, the
bridge checks if that camera is active and sets it as the
active camera if so. It attempts to remove the camera if
active is false (the camera will be removed if it is the point
of view at that point; if it's not, it's a no-op).

3. In order to get this to work on Android and iOS, the
module had to be extracted into a separate CameraModule.
On Android, we cannot combine View managers with modules
like we can on iOS.


Former-commit-id: fa0a6f3d5556f2cd9572b0e48b2741e0332cf8ce