VIRO-765: Resolve Android Renderer Threading Issues (Bridge)
- Dispatched GL specific Viro components to be modified on the Renderer Thread.
These are mainly node specific components that are examined / modified in the
render pass.
- Added ability to post onto Android's UI Applicaiton thread via VROPlatformUtil,
where delegate events to be sent back to JS are then posted on the UI thread.
- Modified VideoTextures to be created asynchronously to account for gl calls.
- Moved the direct calling into Native for Camera position to be async instead.
- Moved the check of containsChild from NodeJNI into the Node view layer so
as to remove the requirement of calling into Native; we use the view hiearchy
in android to perform the same check instead.


Former-commit-id: 9d31395a3ecf89042f2217887beeafb1d66649e1