VIRO-827 [bridge]: material animations occur once
Material animations were not running more than
once because we were mistakenly using the material
in the MaterialManager directly in the geometry,
instead of copying it. In the example in this bug, this
means the first time we ran the animation, we were
replacing the "hover off" material with the "hover on"
texture; so when we went back to the "hover off"
material, the texture wouldn't change.

Fix by copying materials, as we should have been doing
previously. For Android the change is required in the
renderer, for iOS in the bridge.


Former-commit-id: d7ccdf4271fbd10c0a45e3ba03e53935bd85a9c7