VIRO-877: introduce viewWillAppear and viewWillDisappear
Views were only getting events when their parent
scene was inserted or removed. This commit does
the following:

1. Changes the existing viewWillAppear and
viewWillDisappear methods to sceneWillAppear
and sceneWillDisappear, as this more accurately
reflects when those methods are invoked.

2. Introduces *new* willViewAppear and viewWillDisappear
methods that are invoked when a VRTView is added or
removed from the react tree.

3. Uses this change to make sounds stop when the views
are removed from the tree.


Former-commit-id: 02b3e65b0b069163030144f45210b5f0c4966984