VIRO-363,VIRO-834: Video and Sound keeps playing on Reload, App pause and new Scene
- On Android, added React's pause and resume callbacks to our Viro components
to be notified of when the user leaves / re-enters the app. Video and Sound
then pauses or resumes it's players as a result of these callbacks. In addition,
these events are also triggered through reloading, and so resolves that issue
as well.
- Fixed Scene Callback issues as they were not getting triggered due to a weak
reference in the renderer. Tied scene lifecycle events to also pause / resume
video.
- On iOS, removeFromSuperview is called when a reload occurs and the rootView
removes VRTScenenNav. Thus, also cascade viewWillDisappear callbacks to the
tree as well, where Video and Sound can perform cleanup.


Former-commit-id: 29d16a7242ecc08072c7c97cbf8f68ca2460a591