VIRO-919: crash when launching if views removed
When launching the TestBed app, if you're in portrait
mode 2 views get added (the VR view and the first
scene). Then the app switches to landscape so those
two views get removed, and added again. In this
interval, if the render context is set, we get a crash.

Add an additional check to ensure this doesn't happen.

Why the views are added and removed is still an unknown,
but this commit is meant primarily unblock the testbed app.


Former-commit-id: 83a05b11e5ef8c3b3f5a5e522276fac9115a487f