VIRO-799: button tap and gaze source fixes
If tap or gaze source was missing we would get an
error. Fix this so instead if gaze source is missing
we just use the normal image for gaze. If tap source
is missing we use the gaze source, and if there's
no gaze source, we use the normal source.


Former-commit-id: b18b946d02ecec6c8409b20c0803794052cef431