VIRO-947: onTearDown when dropped and detached
Fix onTearDown so that we only tear views down when
they are both dropped and detached (meaning they have
no children and no parent).

Tested and this seems to work: complex view trees are
torn down, but only after React is done processing them.

Still crashes on animations, which continue to run on
town-down components.


Former-commit-id: 9c4b3298c941f895e934227be8ebde08d8c336a1