VIRO-922 [memory]: cascade tear down on reload
Unfortunately React Android does not provide any
direct method to know when it's done with a view
upon reload. However, investigation revealed that
onDetachedFromWindow can be used for this
purpose: on a SceneNavigator it's only invoked when
we reload, or when we remove the SceneNavigator
entirely.

Upon receiving this event we can therefore cascade
down the view tree forcing everything to tear down.

Furthermore, when receiving this event we need to
get the renderer to entirely shut down and delete itself.
(This is because, after a reload, an entirely new
SceneNavigator, ViroGvrLayout, and renderer are
created). We destroy the previous renderer via the
new mVrView.destroy() method.


Former-commit-id: c02643ae4ea8575bf878e1527f551ce4b505a940