VIRO-927, VIRO-928: Exit VR for Hybrid / Non Hybrid React Viro Applications
- Modified SceneNavigator js to allow developers to listen for onViroExit events.
These events can be triggered when the user presses "x" within GVR, or if the
developer specifically exits the vr experience by calling exitViro within the
Scene Navigator.

- Enables developers to pop the react route containing Viro within JS for
ReactJS to Viro JS applications.

- Enables developers to listen for notifcation events for Native to Viro JS
applications, so as to finish the activity / view controller. Specifically
for a ON_EXIT_VIRO_BROADCAST within Android, and kVRTOnExitViro notification in iOS.

- Added in VRTSceneNavigator additional cleanup code as GVR does not perform
proper cleanup of it's windows in iOS.


Former-commit-id: 3ec3a52e9f813378a8bf79c8b7e6a57cb0290263