VIRO-411: on iOS do not download placeholder
On iOS it takes time to download the placeholder
image from the package server, resulting in FPS
drops in development mode, because placeholder
downloads are done synchronously. With this
commit:

1. If no placeholder is provided, do not use the
default viro_blank png. Instead set the material
transparency of the image view to 0.0 until the
image is downloaded. This should have the same
effect as the viro_blank method.

2. Do not change Android, instead detect iOS
and do this only on iOS.

3. Remove VRTUIImageWrapper from VRTImage
on iOS. We can just use RCTConvert’s normal
UIImage. The only reason we were using
the VRTUIImageWrapper was to enable downloading
remote sources synchronously, which is bad. We
still do this for materials but that’s filed as a bug
in JIRA.


Former-commit-id: 189cc317e389c41cd77bb0098f4f9549a414e979