VIRO-1314: refreshable animations + Android support
Adds skeletal animations to Android, using the new changes
in the renderer. This works identically to iOS.

Also made it possible to refresh ManagedAnimations. Each
time an animation is started, it tries to reload the animation.
The ManagedAnimation subclass can then load the animation
again from its sources. This allows developers to change
the properties of an animation and have those new properties
reflected the next time the animation is run.

Tested for identical behavior on iOS and Android using the
ViroAnimatedComponentTest and the Viro3DObjectTest.


Former-commit-id: dcdbb38b0a2ae4fe8a754e2c6b3cd36baf231ca3