VIRO-1729: onClick on scene is not triggered
The issue was that the scene's VRONode as referred to in the renderer, is
not the same node as referred to in the bridge VRTScene - we were actually
creating an additional node and then assigning all out event callbacks on that,
hence we do not get click callbacks at the root level.

The fix is to pass on the VROSceneController's scene vroNode when constrcuting
the VRTScene, this way, both references stays in sync.


Former-commit-id: 5e3270b7631cf23df9f9a13ab0f03cb07bf0fc05