VIRO-1953: [bridge] use need immutable material constructor
Changing materials to be set on the rendering thread
broke parts of the bridge, because material properties
were getting set on the bridge *after* those materials
were copied to their respective components. Instead,
the bridge now uses the immutable constructor here, which
sets all properties on the calling thread (e.g. the UI thread).
With all material parameters provided up-front, the
materials are immediately ready to be assigned to components,
and do not need to call the rendering thread dispatch
setters.


Former-commit-id: 179b9649b14cab1e098ca61acd77c3105b3619a8