VIRO-2486: do not force landscape in 360 mode (bridge)
Change VRTSceneNavigator so it can use the new
forceLandscape boolean on VROViewControllerGVR
to make it such that we do *not* force landscape
mode when using 360.

Unfortunately because the VROViewControllerGVR
needs this information right after construction, we
have to defer the creation of the
VROViewControllerGVR until setVrMode has been
called. To ensure this always works, we lazily load
the VROViewControllerGVR using a new method.


Former-commit-id: 9402fef2c9f99fb23815e50931574c4fc6feaffc