Have Separate React Material fields for PBR Textures and Values
In the ReactBridge, pbr textures / maps for metal and roughness can also be resolved
into a numeric value reference for the bridge to download said asset. However, this
conflicts with actual metal and roughness floats that can also be set on a given
material. As such, the solution is to have roughness and metalness represent the floating
values of a pbr material, and roughnessTexture, metalnessTexture and ambientOcclusionTexture
represent the mapping values of a pbr material.


Former-commit-id: 12fb28f8edf6106464ab8bde8457d3955135787e