VIRO-2407: bridge resonance audio support
Update the bridge to use the new renderer sound
APIs. On Android this was straightforward, eliminating
the local flag in favor of the new resource type methods.

On iOS there were a number of minor issues preventing
spatial sounds from working. Resolved all but one of these
in this commit. The one remaining issue is the VROBaseSound
shouldAppear check prevents spatial sounds from playing.


Former-commit-id: 3c61953cd5e848a975974bf1f543b2bbb7028709