VIRO-2970, VIRO-2974: 360 Photos Show Black
- A release build of react native strips all file extensions for assets and
resources, including the needed .hdr extension for determining hdr download
compatiblity, resulting in an incorrectly downloaded hdr file. This issue
is similiar to problem we had with 3D model file extensions. As such, the
solution is to also provide the ability to specify if the 360 image is hdr
or not within VRT360Image.


Former-commit-id: ca99b91fe6b1f0aa070072e88d102e73f8087966