VIRO-3474 Materials not available after reloading
- In older React Native version, when the user reloaded, RN would create
new view managers as well as modules and use those new ViewManagers and
Modules to create the scene. In the latest version, they only create new
Modules and this results in the old managers attempting to loading materials
from the old MaterialModule thus throwing an error because the materials
can't be found.

The solution is to make the material, animation and target maps static and
shared between all instances of the modules (so the when the new ones create
the objects, the old modules can access them).


Former-commit-id: ce869379da868b895c132ecb07f3cf1bd9e3df11