Properly inheirit exported events
- the two important managers are VRTNodeManager and VRTViroViewGroupManager which
every component inheirits from. So, if one is to add an event to VRTNodeManager, then
either every component needs to have a prop for that or a "nativeOnly" prop must be
added. We were doing this incorrectly before (but getting away with it) because as long
as ANY component exported an event, then any other component could use that event without
erroring out. This was okay until you used a component in isolation of the other components
which then exposed the lack of exporting (see VIRO-2771)


Former-commit-id: 53e8113138ed128435864fe7537e472e670a836e